moving config into plugins
